LPS-106315 Move V10aOAuth and V10aOAuthDebugCacheListener from api-module to service-module.
